+++ R/spread_se.R
@@ -1,12 +1,12 @@
 
 #' Collect values from blocks of rows into columns.
 #'
-#' Call \code{\link[tidyr]{spread}}.
+#' Standared inteface to \code{\link[tidyr]{spread}}.
 #' Take values from the columns named in the \code{columns} argument and
 #' move them into blocks of rows, placing values in the new column specified by \code{value}
 #' and indicating which column each value came from in the new column specified by \code{key}.
 #'
-#' @param data data.frame to take value from.
+#' @param data data.frame to take values from.
 #' @param key character, name for existing column to get new column names from.
 #' @param value character, name for existing column to take values from.
 #' @param ... not used, force later arguments to bind by name.
@@ -19,12 +19,14 @@
 #' @examples
 #'
 #' d <- wrapr::build_frame(
-#'     'id', 'key' , 'value' |
-#'     1   , 'col1', 'a'     |
-#'     1   , 'col2', '10'    |
-#'     2   , 'col1', 'b'     |
-#'     2   , 'col2', '20'    )
-#' spread_se(d)
+#'     'id', 'name_for_new_column' , 'value_to_take' |
+#'     1   , 'col1'                , 'a'             |
+#'     1   , 'col2'                , '10'            |
+#'     2   , 'col1'                , 'b'             |
+#'     2   , 'col2'                , '20'            )
+#' spread_se(d,
+#'    key = 'name_for_new_column',
+#'    value = 'value_to_take')
 #'
 #' @seealso \code{\link[tidyr]{spread}}, \code{\link{gather_se}}
 #'
@@ -42,9 +44,15 @@ spread_se <- function(data,
   if(!(is.data.frame(data) || dplyr::is.tbl(data))) {
     stop("seplyr::spread_se first argument must be a data.frame or tbl")
   }
+  if((!is.character(key))||(length(key)!=1)) {
+    stop("seplyr::spread_se key must be a single string")
+  }
+  if((!is.character(value))||(length(value)!=1)) {
+    stop("seplyr::spread_se value must be a single string")
+  }
   tidyr::spread(data,
-                key = key,
-                value = value,
+                key = !!key,
+                value = !!value,
                 fill = fill,
                 convert = convert,
                 drop = drop,